Add constructor that allows injecting a Consumer
In working on a change to a library that uses LiKafkaConsumerImpl and I
found myself wanting to add an integration test that tested that
integration. I wanted to do that by providing a mock Consumer<byte[],
byte[]> so that my integration test didn't actually need Kafka, but
unfortunately there wasn't any way to inject a consumer object.

This change just adds a couple of new constructors so that clients can
provide their own Consumer<byte[], byte[]> if necessary.