split to multiple modules, use tag-driven versioning
this codebase contains 2 usable artifats already - the clients, and the test harness base classes, which before this patch were included in a test artifact.
a test artifact is awkward to produce, publish and depend upon so this splits the project up into li-apache-kafka-clients (which contains the clients, as t always has)
and kafka-test-harness, which holds the base classes for the test harness. these classes are useful in starting/stopping embedded ZK and kafka for integration tests.
there's also a 3rd module which uses the test harness to run integration tests for the clients. this 3rd module does not produce any artifact for downstream users.

this change also gets rid of the version property and instead uses a gradle plugin to deduce the version based on the "nearest" git tag